LibWeb: Move screenshot requests from traversables to Page
Screenshot requests belong to the page's rendering surface. Store and
process them on Page so capturing a page no longer requires its root
navigable to be a LocalTraversableNavigable.

Keep using the existing top-level accessor and rendering condition while
moving ownership.
